*fix bug in meanline decorators

+++ libraryKernel/geometryEngine/geoBuilder/bSplineCurve2d_angleDeltaXDeltaYConstructOCC.cpp
@@ -77,7 +77,7 @@ namespace dtOO {
 //        int ss = ( deltaY - deltaX * sin(angleIn) / cos(angleIn) )
 //                 /
 //                 ( sin(angleOut) - (cos(angleOut)*sin(angleIn) / cos(angleIn)) );
-        int tt = ( deltaY - deltaX * sin(angleOut) / cos(angleOut) )
+        float tt = ( deltaY - deltaX * sin(angleOut) / cos(angleOut) )
                  /
                  ( sin(angleIn) - (cos(angleIn)*sin(angleOut) / cos(angleOut)) );      
 //        if (!mirrorY) {
